# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.RsError:
    properties:
      code:
        type: integer
      detail:
        type: string
      message:
        type: string
    type: object
...
  subscription.RqUpdatePractitionerSubscription:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/subscription.Status'
        enum:
        - ACTIVE
        - PAST_DUE
        - CANCELLED
        - PAUSED
        - EXPIRED
    type: object
...
  subscription.RsPractitionerSubscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: integer
      practitioner_id:
        type: string
      start_date:
        type: string
      status:
        $ref: '#/definitions/subscription.Status'
      subscription_id:
        type: integer
      updated_at:
        type: string
    type: object
  subscription.RsSubscription:
    properties:
      created_at:
        type: string
      description:
        type: string
      duration_days:
        type: integer
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
  subscription.Status:
    enum:
    - ACTIVE
    - PAST_DUE
    - CANCELLED
    - PAUSED
    - EXPIRED
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusPastDue
    - StatusCancelled
    - StatusPaused
    - StatusExpired
  util.RsList:
    properties:
      items: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Backend API for acareca
  title: Backend API
  version: 1.0.0
paths:
...
  /calculate/{id}:
    get:
      description: Calculate results for a specific form by ID
      parameters:
      - description: Form ID
        in: path
        name: id
        required: true
        type: string
      - description: Super component value override
        in: query
        name: super_component
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Run calculation for a form
      tags:
      - calculation
...
  /practitioner/subscription:
    get:
      consumes:
      - application/json
      description: list subscriptions by practitioner ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.RsList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: List subscriptions by practitioner ID
      tags:
      - subscription
    post:
      consumes:
      - application/json
      description: create a subscription for a practitioner
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/subscription.RsPractitionerSubscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Create a subscription
      tags:
      - subscription
  /practitioner/subscription/{sub_id}:
    delete:
      consumes:
      - application/json
      description: delete a subscription
      parameters:
      - description: Subscription ID
        in: path
        name: sub_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Delete a subscription
      tags:
      - subscription
    get:
      consumes:
      - application/json
      description: get a subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: sub_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.RsSubscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Get a subscription by ID
      tags:
      - subscription
    patch:
      consumes:
      - application/json
      description: update a subscription
      parameters:
      - description: Subscription ID
        in: path
        name: sub_id
        required: true
        type: integer
      - description: Updated Subscription Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/subscription.RqUpdatePractitionerSubscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/subscription.RsSubscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.RsError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.RsError'
      security:
      - BearerToken: []
      summary: Update a subscription
      tags:
      - subscription
...
schemes:
- http
- https
securityDefinitions:
  BearerToken:
    description: Type "Bearer <your_token>" to authenticate
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"